Refactor app to use api folder.
