+++ api/classes/member_show.py
@@ -1,4 +1,4 @@
-import modules.formattable as format
+import api.modules.formattable as format
 
 class MemberShow:
     def __init__(self, data):
