+++ api/models/user.py
@@ -1,5 +1,6 @@
 from api.app import db
 
+
 class User(db.Model):
     __tablename__ = 'users'
